+++ packages/server/lib/socket-base.ts
@@ -45,12 +45,12 @@ const runnerEvents: RunnerEvent[] = [
 
 type ReporterEvent =
   'runner:restart'
- | 'runner:abort'
- | 'runner:console:log'
- | 'runner:console:error'
- | 'runner:show:snapshot'
- | 'runner:hide:snapshot'
- | 'reporter:restarted'
+  | 'runner:abort'
+  | 'runner:console:log'
+  | 'runner:console:error'
+  | 'runner:show:snapshot'
+  | 'runner:hide:snapshot'
+  | 'reporter:restarted'
 
 const reporterEvents: ReporterEvent[] = [
   // "go:to:file"