+++ system-tests/lib/fixtures.ts
@@ -1,15 +1,15 @@
 import fs from 'fs-extra'
 import _path from 'path'
 import chokidar from 'chokidar'
-import os from 'os'
 import cachedir from 'cachedir'
 import execa from 'execa'
+import tempDir from 'temp-dir'
 
 const root = _path.join(__dirname, '..')
 
 const serverRoot = _path.join(__dirname, '../../packages/server/')
 const projects = _path.join(root, 'projects')
-const tmpDir = _path.join(os.tmpdir(), 'cy-projects')
+const cyTmpDir = _path.join(tempDir, 'cy-projects')
 
 // copy contents instead of deleting+creating new file, which can cause
 // filewatchers to lose track of toFile.
@@ -27,9 +27,9 @@ const copyContents = (fromFile, toFile) => {
 }
 
 // copies all of the project fixtures
-// to the tmpDir .projects in the root
+// to the cyTmpDir .projects in the root
 export function scaffold () {
-  fs.copySync(projects, tmpDir)
+  fs.copySync(projects, cyTmpDir)
 }
 
 /**
@@ -37,7 +37,7 @@ export function scaffold () {
  */
 export function scaffoldProject (project: string): void {
   const from = _path.join(projects, project)
-  const to = _path.join(tmpDir, project)
+  const to = _path.join(cyTmpDir, project)
 
   fs.copySync(from, to)
 }
@@ -101,18 +101,16 @@ async function makeWorkspacePackagesAbsolute (pathToPkgJson: string): Promise<st
 }
 
 function getYarnCommand (opts: {
-  yarnV2: boolean
+  yarnV311: boolean
   updateYarnLock: boolean
   isCI: boolean
   runScripts: boolean
 }): string {
   let cmd = `yarn install`
 
-  if (opts.yarnV2) {
-    // yarn v2's docs are no longer available on their site now that yarn v3 is out,
-    // Internet Archive has them here:
-    // @see https://web.archive.org/web/20210102223647/https://yarnpkg.com/cli/install
-    if (!opts.runScripts) cmd += ' --skip-builds'
+  if (opts.yarnV311) {
+    // @see https://yarnpkg.com/cli/install
+    if (!opts.runScripts) cmd += ' --mode=skip-build'
 
     if (!opts.updateYarnLock) cmd += ' --immutable'
 
@@ -131,7 +129,7 @@ function getYarnCommand (opts: {
 
   // in CircleCI, this offline cache can be used
   if (opts.isCI) cmd += ` --cache-folder=~/.yarn-${process.platform} `
-  else cmd += ` --cache-folder=${_path.join(os.tmpdir(), 'cy-system-tests-yarn-cache', String(Date.now()))}`
+  else cmd += ` --cache-folder=${_path.join(tempDir, 'cy-system-tests-yarn-cache', String(Date.now()))}`
 
   return cmd
 }
@@ -150,7 +148,7 @@ type SystemTestPkgJson = {
   /**
    * Run the yarn v2-style install command instead of yarn v1-style.
    */
-  _cyYarnV2?: boolean
+  _cyYarnV311?: boolean
   /**
    * By default, the automatic `yarn install` will not run postinstall scripts. This
    * option, if set, will cause postinstall scripts to run for this project.
@@ -230,7 +228,7 @@ export async function scaffoldProjectNodeModules (project: string, updateYarnLoc
     // 4. Run `yarn install`.
     const cmd = getYarnCommand({
       updateYarnLock,
-      yarnV2: pkgJson._cyYarnV2,
+      yarnV311: pkgJson._cyYarnV311,
       isCI: !!process.env.CI,
       runScripts: pkgJson._cyRunScripts,
     })
@@ -291,7 +289,7 @@ export async function scaffoldCommonNodeModules () {
 }
 
 export async function symlinkNodeModule (pkg) {
-  const from = _path.join(tmpDir, 'node_modules', pkg)
+  const from = _path.join(cyTmpDir, 'node_modules', pkg)
   const to = pathToPackage(pkg)
 
   await fs.ensureDir(_path.dirname(from))
@@ -312,7 +310,7 @@ export function scaffoldWatch () {
   chokidar.watch(watchdir, {
   })
   .on('change', (srcFilepath, stats) => {
-    const tmpFilepath = _path.join(tmpDir, _path.relative(watchdir, srcFilepath))
+    const tmpFilepath = _path.join(cyTmpDir, _path.relative(watchdir, srcFilepath))
 
     return copyContents(srcFilepath, tmpFilepath)
   })
@@ -320,19 +318,19 @@ export function scaffoldWatch () {
 }
 
 // removes all of the project fixtures
-// from the tmpDir .projects in the root
+// from the cyTmpDir .projects in the root
 export function remove () {
-  return fs.removeSync(tmpDir)
+  return fs.removeSync(cyTmpDir)
 }
 
 // returns the path to project fixture
-// in the tmpDir
+// in the cyTmpDir
 export function project (...args) {
   return this.projectPath.apply(this, args)
 }
 
 export function projectPath (name) {
-  return _path.join(tmpDir, name)
+  return _path.join(cyTmpDir, name)
 }
 
 export function get (fixture, encoding: BufferEncoding = 'utf8') {